refactor: extract shared gh CLI token resolution
github-rest.ts's ghAuthToken/isBearerCredentialSyntax and sync-tracker.ts's
keychainCredential independently re-implemented "call gh auth token, strip
trailing newline, validate Bearer syntax". Unify into utils/gh-cli.ts.
sync-tracker.ts now also inherits the GITHUB_TOKEN env-stripping before
shelling out to gh, matching its own doc comment that this reads the
keychain credential rather than an inherited env var.
